+++ src/main/java/blackjack/controller/BlackjackController.java
@@ -3,23 +3,18 @@ package blackjack.controller;
 import blackjack.domain.card.CardDeck;
 import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.Player;
+import blackjack.domain.participant.Players;
 import blackjack.domain.result.GameResult;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 public class BlackjackController {
-
-    public static final String SAME_NAME_ERROR = "중복된 이름을 사용할 수 없습니다.";
-
     public void run() {
         final CardDeck cardDeck = new CardDeck();
         final Dealer dealer = new Dealer();
-        final List<Player> players = playerSetUp();
+        final Players players = playerSetUp();
 
         distributeInitialCards(players, dealer, cardDeck);
         playerGameProgress(players, cardDeck);
@@ -27,40 +22,39 @@ public class BlackjackController {
         showGameResult(players, dealer);
     }
 
-    private List<Player> playerSetUp() {
+    private Players playerSetUp() {
         final List<String> names = InputView.requestName();
         try {
-            validateSameName(names);
-            return generatePlayers(names);
+            return new Players(names);
         } catch (IllegalArgumentException e) {
             OutputView.showErrorMessage(e.getMessage());
             return playerSetUp();
         }
     }
 
-    private void validateSameName(final List<String> names) {
-        final Set<String> changedNames = new HashSet<>(names);
-        if (names.size() != changedNames.size()) {
-            throw new IllegalArgumentException(SAME_NAME_ERROR);
-        }
-    }
+//    private void validateSameName(final List<String> names) {
+//        final Set<String> changedNames = new HashSet<>(names);
+//        if (names.size() != changedNames.size()) {
+//            throw new IllegalArgumentException(SAME_NAME_ERROR);
+//        }
+//    }
 
-    private List<Player> generatePlayers(final List<String> names) {
-        return names.stream()
-                .map(Player::new)
-                .collect(Collectors.toList());
-    }
+//    private List<Player> generatePlayers(final List<String> names) {
+//        return names.stream()
+//                .map(Player::new)
+//                .collect(Collectors.toList());
+//    }
 
-    private void distributeInitialCards(final List<Player> players, final Dealer dealer, final CardDeck cardDeck) {
-        for (final Player player : players) {
+    private void distributeInitialCards(final Players players, final Dealer dealer, final CardDeck cardDeck) {
+        for (final Player player : players.getPlayers()) {
             player.receiveInitialCards(cardDeck);
         }
         dealer.receiveInitialCards(cardDeck);
-        OutputView.showDistributedCard(players, dealer);
+        OutputView.showDistributedCard(players.getPlayers(), dealer);
     }
 
-    private void playerGameProgress(final List<Player> players, final CardDeck cardDeck) {
-        for (final Player player : players) {
+    private void playerGameProgress(final Players players, final CardDeck cardDeck) {
+        for (final Player player : players.getPlayers()) {
             singlePlayerGameProgress(cardDeck, player);
         }
     }
@@ -84,10 +78,10 @@ public class BlackjackController {
         }
     }
 
-    private void showGameResult(final List<Player> players, final Dealer dealer) {
-        OutputView.showFinalCardResult(players, dealer);
-        final GameResult gameResult = new GameResult(dealer, players);
+    private void showGameResult(final Players players, final Dealer dealer) {
+        OutputView.showFinalCardResult(players.getPlayers(), dealer);
+        final GameResult gameResult = new GameResult(dealer, players.getPlayers());
         OutputView.showGameResult(dealer, gameResult.getResultCounts());
-        OutputView.showPlayersGameResult(players, gameResult.getPlayersResult());
+        OutputView.showPlayersGameResult(players.getPlayers(), gameResult.getPlayersResult());
     }
 }
